+++ src/edwh/local_tasks/plugin.py
@@ -12,8 +12,7 @@ from dataclasses import dataclass
 from typing import Optional
 
 import dateutil.parser
-import requests
-import yayarl
+import yayarl as yarl
 from invoke import Context  # type: ignore
 from packaging.version import parse as parse_package_version
 from termcolor import colored, cprint
@@ -154,8 +153,8 @@ def gather_plugin_info(c: Context) -> list[Plugin]:
     available_plugins = ["edwh", *_get_available_plugins_from_pypi("edwh", "plugins")]
 
     installed_plugins_raw = list_installed_plugins(c)
-    if not installed_plugins_raw or len(installed_plugins_raw) == 1 and installed_plugins_raw[0] == "":
-        raise ModuleNotFoundError("No 'edwh' packages found. That can't be right")
+    if not installed_plugins_raw or (len(installed_plugins_raw) == 1 and installed_plugins_raw[0] == ""):
+        cprint("No 'edwh' packages found. That can't be right", color="yellow")
 
     return _gather_plugin_info(c, available_plugins)
 
@@ -319,17 +318,17 @@ def remove(c: Context, plugin_names: str) -> None:
     c.run(f"{pip} uninstall " + " ".join(plugin_names_splitted))
 
 
-GITHUB_RAW_URL = yayarl.URL("https://raw.githubusercontent.com")
+GITHUB_RAW_URL = yarl.URL("https://raw.githubusercontent.com")
 
 
-def get_changelog(github_repo: str | yayarl.URL) -> str:
+def get_changelog(github_repo: str | yarl.URL) -> str:
     if isinstance(github_repo, str):
-        github_repo = yayarl.URL(github_repo)
+        github_repo = yarl.URL(github_repo)
 
     github_repo = github_repo.path.removeprefix("/")  # e.g. educationwarehouse/edwh
     changelog_url = GITHUB_RAW_URL / github_repo / "master/CHANGELOG.md"  # replace github.com with github raw
 
-    return requests.get(str(changelog_url), timeout=10).text
+    return changelog_url.get(timeout=10).text
 
 
 def get_changelogs_threaded(github_repos: dict[str, str]) -> dict[str, str]: